+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/controllers/FileController.java
@@ -48,6 +48,7 @@ import org.json.simple.parser.JSONParser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -62,6 +63,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.util.UriComponents;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.File;
@@ -70,6 +73,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.InetAddress;
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
@@ -145,6 +149,28 @@ public class FileController {
                 .header(HttpHeaders.CONTENT_DISPOSITION,
                         "attachment; filename=\"" + resource.getFilename() + "\"")
                 .body(resource);
+    }
+
+    private ResponseEntity<Resource> downloadSample(final String fileName) {
+        String serverPath = System.getProperty("user.dir");
+        String contentType = "application/octet-stream";
+        String[] fileLocation = new String[] {serverPath, "src", "main", "resources", "assets", "document-templates",
+                                              "sample", fileName};
+        Path filePath = Paths.get(String.join(File.separator, fileLocation));
+        Resource resource;
+        try {
+            resource = new UrlResource(filePath.toUri());
+            if (resource.exists()) {
+                return ResponseEntity.ok()
+                        .contentType(MediaType.parseMediaType(contentType))
+                        .header(HttpHeaders.CONTENT_DISPOSITION,
+                                "attachment; filename=\"" + resource.getFilename() + "\"")
+                        .body(resource);
+        }
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+        }
+        return null;
     }
 
     // download data from the specified history file
@@ -372,14 +398,12 @@ public class FileController {
     @GetMapping("/assets")
     public ResponseEntity<Resource> assets(@RequestParam("name")
                                                final String name) {  // get sample files from the assests
-        String fileName = Path.of("assets", "document-templates", "sample", fileUtility.getFileName(name)).toString();
-        return downloadFile(fileName);
+        return downloadSample(name);
     }
 
     @GetMapping("/csv")
     public ResponseEntity<Resource> csv() {  // download a csv file
-        String fileName = Path.of("assets", "document-templates", "sample", "csv.csv").toString();
-        return downloadFile(fileName);
+        return downloadSample("csv.csv");
     }
 
     @GetMapping("/files")
@@ -485,6 +509,22 @@ public class FileController {
                 }
             }
 
+            String link = body.getLink();
+            if (fileName.equals("") && link != null) {
+                if (!link.contains(storagePathBuilder.getServerUrl(true))) {
+                    HashMap<String, String> data = new HashMap<>();
+                    data.put("url", link);
+                    data.put("directUrl", link);
+                    return gson.toJson(data);
+                }
+
+                UriComponents uriComponents = UriComponentsBuilder.fromUriString(body.getLink()).build();
+                fileName = uriComponents.getQueryParams().getFirst("fileName");
+                boolean fileExists = new File(storagePathBuilder.getFileLocation(fileName)).exists();
+                if (!fileExists) {
+                    return "{ \"error\": \"File is not exist\"}";
+                }
+            }
 
             if (fileName.equals("")) {
                 try {